feat: API for adding measurement programatically
